Pass mutable scene builder reference to paint
Instead of the `paint` method returning a `Rendered` which was just a
newtype around `SceneFragment`, pass in `&mut SceneBuilder`. Also store
a scene fragment in each `Pod`, which is the storage for these builders.

This change has a number of consequences. Less boilerplate in the paint
method, and also less allocation, as the storage can get reused. More
importantly, though we don't have the logic for this yet, it sets the
stage for retaining scene fragments when widgets don't change.