Fix accessibility propagation
This patch makes a number of changes to change flags propagation, which should at least get the example working again.

This implements the two-flags approach to REQUEST_ACCESSIBILITY and DESCENDANT_REQUESTED_ACESSIBILITY. The framework converts the former flag into the latter as it propagates up the tree, both in `merge_up` and in ChangeFlags propagation.

ChangeFlags propagation only propagates ChangeFlags, it does not move flags from the pod into the rebuild flow. It does have descendant logic as mentioned above.

Layout now triggers an accessibility request. This reduces the burden on custom widgets to explicity request accessibility, and also will be more precise if and when layout becomes more incremental, we'll want to request accessibility only when the layout has actually changed.