Small refactor of flags to avoid duplication
This changes the logic so the DESCENDANT_REQUESTED_ACCESSIBILITY logic is only written once. It also adds a comment warning that more changes will be needed if we get ChangeFlags that are not also PodFlags.

I think this will happen when there is a ChangeFlags::STRUCTURE. I'm not sure what's going to be the best way to handle it, but we can cross that bridge when that flag is added.