Draft of work to make AnyView generic
This should be considered only a draft. The AnyView module itself compiles, but so far when I've tested it I haven't been able to work.

Also there are some organizational and doc difficulties. The new AsAny trait may not have the best name (maybe BoxedAsAny?) and probably doesn't belong in the widget tree.