Make main traits generic on Send
Add additional arguments to the macros which instantiate view, allowing them optionally to be Send.

The main widget tree is Send because we're pushing the ability to run app logic in multiple threads. A web deployment would not be Send, because many primitives needed for web integration (including closures) are !Send.