Add xilemsvg example
This is a proof of concept that uses the common xilem_core to generate vector graphics content rendered as SVG DOM nodes in the browser.

This could plausibly be made into an actual library, and it's likely that an evolution of this will be used to create interactive illustrations, for example in the raphlinus blog.