Make the HTML elements more typed
Don't store the name in the element view, but use a zero-sized type.

Note, this breaks the untyped case. That will probably need a separate view, or alternatively the tag doesn't have to be zero-sized (so there is a variant that stores a String), the name func could take a &self parameter, and it could return the internal string.

Note, also uses the stringify function to reduce the duplication in the macros. That's a separate, orthogonal change.