Update to current-ish glazier.
The change to `app_main.rs` matches a similar change made in the
Glazier examples.

Also, since we now have both x11 and wayland backends on some
platforms, we need to make sure that we create the application
prior to creating the menu. This will be improved in future
versions of Glazier.

This minimally updates to the PointerEvent API within Glazier
as that has replaced the MouseEvent API.