xilem_html: Add an (ordered) `VecMap` for small `n`
A few benchmarks have shown, that for small n (roughly `n < ~20`, with small types < 50 bytes) a Vec is faster than a BTreeMap.
This seems to be the average case for html attributes.

Since it was a low hanging fruit to boldly copy the docs from the `std::collections::BTreeMap`, I have done so, it also implicitly adds doc tests for this

The VecMap is now used instead of the BTreeMap for the element attributes