xilem_html: Add support for different types in `el.attr("attr", <value>)`  (#127)
Where `<value>` can be primitive values like numbers, booleans and strings, and options of these values.
In case a `None` or `false` value is passed, the attribute is removed from the element, as even an empty string value for an attribute evaluates to `true` in javascript (relevant for e.g. `input.attr("checked", false)`).
This is achieved via an enum that serializes to a string and syntax sugar via an extra trait `IntoAttributeValue` which converts a (currently) primitive value into this enum when passing it to the method.
This also adds an additional method `el.remove_attr(name)`.