masonry: Use workspace lint config.
This replaces the old lint configuration in a cargo config with
the workspace lint configuration in the Cargo.toml.

The old lint configuration was for tests, which don't currently
trigger those lints, so we don't need to supply this configuration
elsewhere. (The library has its own lint configuration that
supplements this in `src/lib.rs` which can be improved later.)