xilem_masonry: Robust generational indexing for `ViewSequence` (#220)
Alternative to #218

The version of viewsequence added in #205 is robust for all properties
supported in that version. However, it would be incompatible with future
expansions to the Xilem model, in particular async wiring.

In this PR, I propose a system to resolve this, which uses generational
indexes in the view id path (for the items where this is relevant).

Some other historical record: #9

This brings back the `View::State` associated type (renamed to
`ViewState` and `SeqState` depending on the trait), in order to perform
this wiring.