masonry: Fix non-resetting visual active state when releasing active button outside of its area (#247)
When pressing the button and releasing it outside of its active area,
the visual pressed state (background gradient) stays "pressed". This
should fix this.

(Not related to that PR) I'm not sure, if there's a case, where the
disabled state could change while pressing the button and trigger a
similar behavior.