xilem_masonry: Fix Flex splice (increment ix by 2 because of spacer widget) (#264)
While adding filters to the TodoMVC example, I produced a crash. Which
seems to originate because there was a spacer added, which also
increments the underlying element child count.
So in case we want to have that spacer, we need to increment by 2
instead...

(Since `mutate` and `delete` skip spacers, because `child_mut` returns
none, and the index is only incremented for concrete children it's not
necessary to change the index incrementation code there)