Convert Masonry `README.md` to the Linebender standard. (#271)
This is a basic skeleton based on the latest Linebender standards.

The installation instructions were stale, based on `druid-shell`. The
example code and roadmap were also stale, not reflective of the recent
barrage of changes. Thus I removed these stale sections.

I removed `CONTRIBUTING.md` because the new readme has a section about
that as standard, and a more thorough `CONTRIBUTING.md` will eventually
live in the repo root.

I moved `ARCHITECTURE.md` and `ROADMAP.md` to a new `doc` sub-directory,
as is standard for Linebender. These are actually also stale, but won't
be immortalized by publishing so changing/deleting them isn't as urgent.