Add `README.md` for `xilem`. (#268)
This is a basic skeleton based on the latest Linebender standards.

I also added a paragraph about the reactive layer's origins.
Plus a section describing the modularity of our stack.

Additional sections describing more of the actual new state of things
are welcome in follow-up PRs, but are not a goal of this PR.