deps: Update `image`, make workspace, turn off default features.
The `image` crate has a number of dependencies which aren't needed
in the core masonry crate, so we can turn off the default image
format support and only enable "png" when used as a dev dependency.

Also, make it a workspace dependency at the same time so that the
version is only stated in a single location.