xilem,masonry: Use `dpi` crate directly, re-export it
This removes a class of dependencies from within both `masonry`
and `xilem` on the `winit` crate where we can just use `dpi`
directly instead.

The `dpi` crate is meant (like `cursor_icon`) to be a shared ecosystem
crate rather than only for usage with `winit`.

Both `masonry` and `xilem` now re-export the `dpi` crate for the
convenience of crate consumers.