lints: Suppress some check-cfg warnings (#321)
The current nightly issues warnings for unknown cfgs, so we need to
suppress some of these for now.

Even though these are only needed for `masonry`, we put them in the
workspace configuration due to limitations in how the lints table is
configured and overridden in current versions of Cargo.