masonry: Avoid reconstructing LogicalPosition.
In the call to updating the hot state from the widget pod, the
`mouse_pos` is already of the right type, so it doesn't need
to try to reconstruct itself as the right type.

This is different from another call site (in `contexts.rs`)
where the `mouse_pos` is an `Option<Point>` and so it does
need reconstructing.