Add diagram to document xilem layers (#322)
Note that this adds a link to a section added in
https://github.com/linebender/parley/pull/69

I removed Taffy from the diagram for simplicity, and because Taffy's
current integration is in flux.

Overall I think the Xilem README is in desperate need of a rewrite, and
this PR should be considered a stopgap. Further PRs should better
integrate the ecosystem descriptions into the README, move some sections
around, etc.