Differentiate pointer buttons in `button` widget
In most cases, you want a button which only actuates when the primary mouse button is pressed
This is a near term hack, because e.g. the active state is still based on any button being pressed, not just those
we are interested in.

That is, we need to represent a set of buttons we want to receive
However, this change minimally unblocks additional work with Xilem
In particular, see
[#xilem > Minesweeper converted from Iced to Xilem](https://xi.zulipchat.com/#narrow/stream/354396-xilem/topic/Minesweeper.20converted.20from.20Iced.20to.20Xilem)