xilem_web: Add hydration support, and a new view using it `AsTemplate`
`AsTemplate` stores a deep copy of the DOM node of the first occurence of its `impl DomView` based on the `TypeId` in the `ViewCtx`,
and reuses it on every further invocation, where it will be hydrated and updated based on the new attributes.
As it uses an `Rc` to achieve this, it also supports memoization, similar to an `Arc<impl View>`.

Hydration is currently feature-gated, as it produces a little bit more binary bloat. Though it's little enough to this being justified as default.