Add public methods to access clip path
This adds public getters for clip path to non-`LayoutCtx` contexts, and
also to `WidgetState` as it seems relevant for painting.

The motivation is to remove private field accesses from
`Widget::get_child_at_pos` as mentioned in
https://github.com/linebender/xilem/pull/565#issuecomment-2374162416.

This leaves one private field access, namely `WidgetState::is_stashed`.
As per the docs of `WidgetState` ("widgets will generally not interact
with it directly"), that field perhaps should not be publicly accessible
from `WidgetState`, but then it should be accessible from `WidgetRef`.