Make selection non-optional in TextWithSelection (#628)
There was no good reason for selection to be optional, except to not
display it when focus was lost. This is now handled with a boolean flag.
A side-effect of this is that TextWithSelection now remembers its
selection when focus is gained back.

Selection was used with a default value of `(0, 0)` in cases where it
was `None`, so its default value in similar cases is still `(0, 0)`.