Use standard cargo parameter order in CI. (#722)
This PR swaps the order of the `--each-feature` and `--optional-deps`
flags.

This is the order in the newer standard, as seen in e.g.
[peniko](https://github.com/linebender/peniko/blob/7940e7de0aa09673f522a8386d91cd081af98e87/.github/workflows/ci.yml#L114):
`--optional-deps --each-feature --features std`.

This helps with the goal of keeping all the CI scripts as identical as
possible.

---

This PR also adds some missing comments compared to
[vello](https://github.com/linebender/vello/blob/4c6f75f499a225c4a97df71ade47b0ccc3a5096b/.github/workflows/ci.yml#L42-L44)
and removes a trailing slash to match the [style of
vello](https://github.com/linebender/vello/blob/4c6f75f499a225c4a97df71ade47b0ccc3a5096b/.github/workflows/ci.yml#L247-L249)
as well.