Fix `unused_qualifications` in Xilem Core (#743)
Most of the change is remove the `expect` and run `cargo fix`.

For the orphan view, I've added a blanket exception. At the moment, this
only applies to String, but the

I did try using `#![no_implicit_prelude]` in that module, but there are
a lot of useful things in the implicit prelude, it was worse than the
exception.