Re-add IME support (#762)
This adds IME support back into Masonry. This sticks close to
https://github.com/linebender/parley/pull/111, except that during IME
compose, this version doesn't allow changing the selection anchor,
making the code simpler. For reference, there's also
https://github.com/linebender/parley/pull/136.

This tweaks the focus update pass: when a widget with IME is unfocused,
Masonry sends the platform's IME disable event to the newly focused
widget. As a workaround, we synthesize an IME disable event in the focus
pass and send it to the widget that is about to be unfocused. A
complication is that the handling of that event can request focus to a
different widget, and in particular, can request itself to be focused
again. This handles that case, too.

Remaining work is setting the IME candidate region to be near the
current selection and to make a decision on cursor/selection hiding when
the platform sends a `None` cursor.

---------

Co-authored-by: Daniel McNab <36049421+DJMcNab@users.noreply.github.com>