Clear compose in TextArea::reset_text
The user may be composing when the `TextArea` text is programmatically
changed. This change clears the compose state in order to adhere to the
`PlainEditor` requirements. It is disruptive, but we already warn about
disruption (and about IME in particular) in the documentation of
`TextArea::reset_text`.

This doesn't reset the platform's IME state, we could do, but that's
potentially even more disruptive, and looks like it's a bit harder to
wire up. We could also re-insert the preedit text. That'd be nicest to
do if we can query `PlainEditor` for the current preedit text, which
requires an API change there.